finalizing connection backend <-> frontend

+++ express-backend/src/supabase/db/delete.ts
@@ -1,8 +1,8 @@
 import { deleteFile } from "./files";
 import { prisma } from "../supabase";
 
-const deleteSelectedUrl = async (id: string, qrPath: string) => {
-  await deleteFile(qrPath);
+const deleteSelectedUrl = async (id: string, qrPath?: string) => {
+  if (qrPath) await deleteFile(qrPath);
   const data = await prisma.urls.delete({
     where: {
       id,
